requirements.txt
# requirements.txt for Streamlit Binding Affinity App

streamlit
pandas
numpy
scikit-learn
xgboost
joblib==1.1.0
matplotlib
plotly
graphviz
shap
py3Dmol
biopython
rdkit
lightgbm
catboost
scikit-learn